requirements.txt
Image
pillow
pytesseract
